Added validation sources JSON
Added a JSON file which stores the citations, the number of times it occured, and whether it was a factual source. For each time that it has a source, it checks in the json file whether it was already found, if its not found, it will check PubMed DB, else it will base it on the JSON file. Also restructured the markdown, citations are colored either green or red, based on validity